Animation Pipeline Milestone
This commit represents an effort to establish a rudimentary animation
scheme for the game. Using the Aseprite export_spritesheet function, the
game now can read in animations following a naming convention, and then
make those available to animation_instances in the code: namely, the
player's ship and thruster animations in this commit.

Animations are named with the form <group>_<pixelsize>px_<name>
- 'group' is the name for the animation group, usually related to a game
  entity
- 'pixelsize' should represent the square size of the sprite
- 'name' is anything descriptive

The animations are loaded by consulting the corresponding .json for the
aseprite file. These json's are generated by the export_spritesheet
option in Aseprite, and must be exported as HASH with frame titles of
the template '{frame}'. For now the json, aseprite, and png files should
all live in the 'sprites' directory.

Animations play until they reach their last frame, then maintain the
last frame. If the animation definition specifies loop=true, then it
will instead proceed to the first frame and continue. If the animation
definition specifies next=name, the animation will change to the
animation of the current animation group of the name specified, on its
first frame. This is all configured at the animation loader
initialization.

To add a new animation, make sure it is in the animation loader
initialization hash.

Small Changes include:
- Letterboxing of the game screen (wider window aspect helps show LUA
  errors when things blue screen)
- Added background stars and adjusted laser spawn position