Orbital Scene with Trading Milestone
This commit introduces the planet Isikur in a rudimentary style,
principally making it possible for the player to buy and sell commodity
goods with the planet and make/lose money. This is done via the DEAL
option available in the new OrbitScene, which has other options that are
currently disabled. The DealMenu lets you select a good with the up/down
keys, and to buy/sell the good with the left/right keys. You can
optionally hold down 'z' to buy/sell 5 at a time.

Switching between the shmup CorridorScene and the
OrbitScene is done via the 'n' key to enter orbit, and selecting the
EXIT option to leave orbit.

Minor Changes
- Blinking stars are in the background of both scenes now, scrolling to
  indicate motion in the CorridorScene and staying still in the
  OrbitScene.
- A system font for the game has been decided on for now
- Many things refactored for organization purposes